+++ regression/reference/include/wrapClass2.cpp
@@ -34,9 +34,9 @@ void LIB_Class2_method1(LIB_Class2 * self, MPI_Fint comm)
 // Requested: c
 // Match:     c_default
 // ----------------------------------------
-// Argument:  three::Class1 * c2 +intent(in)
-// Requested: c_shadow_*_in
-// Match:     c_shadow_in
+// Argument:  three::Class1 * c2 +intent(inout)
+// Requested: c_shadow_*_inout
+// Match:     c_shadow_inout
 void LIB_Class2_method2(LIB_Class2 * self, LIB_three_Class1 * c2)
 {
     Class2 *SH_this = static_cast<Class2 *>(self->addr);